+++ ldaptor/protocols/ldap/ldapclient.py
@@ -35,6 +35,7 @@ class LDAPStartTLSBusyError(ldaperrors.LDAPOperationsError):
     def __str__(self):
         return 'Cannot STARTTLS while operations on wire: %r' % self.onwire
 
+
 class LDAPStartTLSInvalidResponseName(ldaperrors.LDAPException):
     def __init__(self, responseName):
         self.responseName = responseName
@@ -42,6 +43,7 @@ class LDAPStartTLSInvalidResponseName(ldaperrors.LDAPException):
 
     def __str__(self):
         return 'Invalid responseName in STARTTLS response: %r' % (self.responseName, )
+
 
 class LDAPClient(protocol.Protocol):
     """An LDAP client"""
@@ -54,8 +56,8 @@ class LDAPClient(protocol.Protocol):
 
     berdecoder = pureldap.LDAPBERDecoderContext_TopLevel(
         inherit=pureldap.LDAPBERDecoderContext_LDAPMessage(
-        fallback=pureldap.LDAPBERDecoderContext(fallback=pureber.BERDecoderContext()),
-        inherit=pureldap.LDAPBERDecoderContext(fallback=pureber.BERDecoderContext())))
+            fallback=pureldap.LDAPBERDecoderContext(fallback=pureber.BERDecoderContext()),
+            inherit=pureldap.LDAPBERDecoderContext(fallback=pureber.BERDecoderContext())))
 
     def dataReceived(self, recd):
         self.buffer += recd
@@ -79,13 +81,13 @@ class LDAPClient(protocol.Protocol):
         # notify handlers of operations in flight
         while self.onwire:
             k, v = self.onwire.popitem()
-            d, _, _, _ = v
+            d, _, _, _, _ = v
             d.errback(reason)
 
-    def _send(self, op):
+    def _send(self, op, controls=None):
         if not self.connected:
             raise LDAPClientConnectionLostException()
-        msg = pureldap.LDAPMessage(op)
+        msg = pureldap.LDAPMessage(op, controls=controls)
         if self.debug:
             log.msg('C->S %s' % repr(msg))
         assert msg.id not in self.onwire
@@ -95,22 +97,20 @@ class LDAPClient(protocol.Protocol):
         d.callback(msg)
         return True
 
-    def send(self, op):
+    def send(self, op, controls=None):
         """
         Send an LDAP operation to the server.
-
         @param op: the operation to send
-
         @type op: LDAPProtocolRequest
-
+        @param controls: Any controls to be included in the request.
+        @type controls: LDAPControls
         @return: the response from server
-
         @rtype: Deferred LDAPProtocolResponse
         """
-        msg = self._send(op)
+        msg = self._send(op, controls=controls)
         assert op.needs_answer
         d = defer.Deferred()
-        self.onwire[msg.id] = (d, None, None, None)
+        self.onwire[msg.id] = (d, False, None, None, None)
         self.transport.write(str(msg))
         return d
 
@@ -119,31 +119,65 @@ class LDAPClient(protocol.Protocol):
         Send an LDAP operation to the server, expecting one or more
         responses.
 
+        If `handler` is provided, it will receive a LDAP response as
+        its first argument. The Deferred returned by this function will
+        never fire.
+
+        If `handler` is not provided, the Deferred returned by this
+        function will fire with the final LDAP response.
+
         @param op: the operation to send
-
         @type op: LDAPProtocolRequest
-
         @param handler: a callable that will be called for each
         response. It should return a boolean, whether this was the
         final response.
-
         @param args: positional arguments to pass to handler
-
         @param kwargs: keyword arguments to pass to handler
-
-        @return: the result from the last handler as a deferred that
-        completes when the last response has been received
-
+        @return: the result from the first handler as a deferred that
+        completes when the first response has been received
         @rtype: Deferred LDAPProtocolResponse
         """
         msg = self._send(op)
         assert op.needs_answer
         d = defer.Deferred()
-        self.onwire[msg.id] = (d, handler, args, kwargs)
+        self.onwire[msg.id] = (d, False, handler, args, kwargs)
         self.transport.write(str(msg))
         return d
 
-    def send_noResponse(self, op):
+    def send_multiResponse_ex(self, op, controls=None, handler=None, *args, **kwargs):
+        """
+        Send an LDAP operation to the server, expecting one or more
+        responses.
+
+        If `handler` is provided, it will receive a LDAP response *and*
+        response controls as its first 2 arguments. The Deferred returned
+        by this function will never fire.
+
+        If `handler` is not provided, the Deferred returned by this
+        function will fire with a tuple of the first LDAP response
+        and any associated response controls.
+
+        @param op: the operation to send
+        @type op: LDAPProtocolRequest
+        @param controls: LDAP controls to send with the message.
+        @type controls: LDAPControls
+        @param handler: a callable that will be called for each
+        response. It should return a boolean, whether this was the
+        final response.
+        @param args: positional arguments to pass to handler
+        @param kwargs: keyword arguments to pass to handler
+        @return: the result from the last handler as a deferred that
+        completes when the last response has been received
+        @rtype: Deferred LDAPProtocolResponse
+        """
+        msg = self._send(op, controls=controls)
+        assert op.needs_answer
+        d = defer.Deferred()
+        self.onwire[msg.id] = (d, True, handler, args, kwargs)
+        self.transport.write(str(msg))
+        return d
+
+    def send_noResponse(self, op, controls=None):
         """
         Send an LDAP operation to the server, with no response
         expected.
@@ -151,7 +185,7 @@ class LDAPClient(protocol.Protocol):
         @param op: the operation to send
         @type op: LDAPProtocolRequest
         """
-        msg = self._send(op)
+        msg = self._send(op, controls=controls)
         assert not op.needs_answer
         self.transport.write(str(msg))
 
@@ -166,21 +200,27 @@ class LDAPClient(protocol.Protocol):
         if msg.id == 0:
             self.unsolicitedNotification(msg.value)
         else:
-            d, handler, args, kwargs = self.onwire[msg.id]
+            d, return_controls, handler, args, kwargs = self.onwire[msg.id]
 
             if handler is None:
-                assert args is None
-                assert kwargs is None
-                d.callback(msg.value)
+                assert (args is None) or (args == ())
+                assert (kwargs is None) or (kwargs == {})
+                if return_controls:
+                    d.callback((msg.value, msg.controls))
+                else:
+                    d.callback(msg.value)
                 del self.onwire[msg.id]
             else:
                 assert args is not None
                 assert kwargs is not None
                 # Return true to mark request as fully handled
-                if handler(msg.value, *args, **kwargs):
-                    del self.onwire[msg.id]
+                if return_controls:
+                    if handler(msg.value, msg.controls, *args, **kwargs):
+                        del self.onwire[msg.id]
+                else:
+                    if handler(msg.value, *args, **kwargs):
+                        del self.onwire[msg.id]
 
-    ##Bind
     def bind(self, dn='', auth=''):
         """
         @depreciated: Use e.bind(auth).
@@ -202,7 +242,6 @@ class LDAPClient(protocol.Protocol):
             raise ldaperrors.get(msg.resultCode, msg.errorMessage)
         return (msg.matchedDN, msg.serverSaslCreds)
 
-    ##Unbind
     def unbind(self):
         if not self.connected:
             raise Exception("Not connected (TODO)")  # TODO make this a real object