+++ docs/source/NEWS.rst
@@ -11,6 +11,8 @@ Features
 - Ability to logically compare ldaptor.protocols.pureldap.LDAPFilter_and and ldaptor.protocols.pureldap.LDAPFilter_or objects with ==.
 - Ability to customize ldaptor.protocols.pureldap.LDAPFilter_* object's encoding of values when using asText.
 - New client recipe- adding an entry to the DIT.
+- Ability to use paged search control for LDAP clients.
+- New client recipie- using the paged search control.
 
 Changes
 ^^^^^^^